Got rid of X,Y,Z FCVariable stuff.  Changes to get rid of g++ warnings.

+++ src/Uintah/Components/Schedulers/MixedScheduler.cc
@@ -1325,8 +1325,8 @@ MixedScheduler::gatherParticles(const ProcessorGroup* pc,
    vector<MPIScatterRecord*> sr;
    vector<int> recvsize(neighbors.size());
    int me = d_myworld->myrank();
-   ASSERTEQ(sgargs.dest.size(), neighbors.size());
-   ASSERTEQ(sgargs.tags.size(), neighbors.size());
+   ASSERTEQ((int)sgargs.dest.size(), (int)neighbors.size());
+   ASSERTEQ((int)sgargs.tags.size(), (int)neighbors.size());
    vector<char*> recvbuf(neighbors.size());
    vector<int> recvpos(neighbors.size());
    for(int i=0;i<(int)neighbors.size();i++){
@@ -1689,6 +1689,9 @@ MixedScheduler::releaseLoadBalancer()
 
 //
 // $Log$
+// Revision 1.8  2000/12/22 00:13:52  jas
+// Got rid of X,Y,Z FCVariable stuff.  Changes to get rid of g++ warnings.
+//
 // Revision 1.7  2000/12/10 09:06:10  sparker
 // Merge from csafe_risky1
 //
